Moved most util from DashArgs class to util folder
