+++ src/structures/DashArgs.js
@@ -1,3 +1,5 @@
+const { identify, sanitize, getKeyVal, typeCoerce } = require('../util');
+
 module.exports = class DashArgs {
     #parsed;
 
@@ -28,89 +30,36 @@ module.exports = class DashArgs {
         return this.#parsed.map(({ key, value, raw }) => ({ key, value, raw }));
     };
 
-    static _typeCoerce(item = '') {
-        if (typeof item != 'string') return item;
-    
-        item = item.trim();
-    
-        if (!isNaN(Number(item))) return Number(item);
-        if (item.toLowerCase() == 'true') return true;
-        if (item.toLowerCase() == 'false') return false;
-        
-        return item;
-    };
-    
     static _parse(string, config) {
         const pattern = /(--[^ \n]+)|(-(?:([^-\s])+)( )?(?:('(?:\.|[^'])*'|"(?:\.|[^"])*")|((?:\.|[^- \n])*)?))/gim;
         let matches = (string.match(pattern) || []).map(m => DashArgs._parseSingle(m, config)).flat();
         if (!config.parseFlags) matches = matches.filter(m => !m.type.match(/flag/gim));
         if (!config.parseArgs) matches = matches.filter(m => !m.type.match(/arg/gim));
-        if (config.typeCoerce) matches = matches.map(m => ({ ...m, value: DashArgs._typeCoerce(m.value) }));
+        if (config.typeCoerce) matches = matches.map(m => ({ ...m, value: typeCoerce(m.value) }));
         return matches;
     };
     
     static _parseSingle(string) {
-        const { key, value } = DashArgs._getKeyVal(string);
-        const { type } = DashArgs._identify(string);
+        const { key, value } = getKeyVal(string);
+        const { type } = identify(string);
         const raw = string;
 
         switch(type) {
             case 'arg':
-                return({ key: DashArgs._sanitize(key, 'key'), value, type, raw });
+                return({ key: sanitize(key, 'key'), value, type, raw });
                 break;
             case 'compound-arg':
-                return({ key: DashArgs._sanitize(key, 'key'), value: DashArgs._sanitize(value, 'value'), type, raw });
+                return({ key: sanitize(key, 'key'), value: sanitize(value, 'value'), type, raw });
                 break;
             case 'flag':
-                if (DashArgs._sanitize(key, 'key').length == 1) return ({ key: DashArgs._sanitize(key, 'key'), value: true, type, raw });
-                return DashArgs._sanitize(key, 'key').split('').map(k => `-${k}`).map(k => DashArgs._parseSingle(k))
+                if (sanitize(key, 'key').length == 1) return ({ key: sanitize(key, 'key'), value: true, type, raw });
+                return sanitize(key, 'key').split('').map(k => `-${k}`).map(k => DashArgs._parseSingle(k))
                 break;
             case 'compound-flag':
-                return({ key: DashArgs._sanitize(key, 'key'), value: true, type, raw })
+                return({ key: sanitize(key, 'key'), value: true, type, raw })
                 break;
             default:
                 return undefined;
-                break;
-        };
-    };
-    
-    static _identify(string) {
-        const { key, value } = DashArgs._getKeyVal(string);
-        let type;
-        if (value) {
-            if (value.match(/(?:^"[^]+"$)|(?:^'[^]+'$)/gim)) {
-                type = 'compound-arg'
-            } else {
-                type = 'arg';
-            };
-        } else if (key && !value) {
-            if (key.match(/(-){2}/g)) {
-                type = 'compound-flag';
-            } else {
-                type = 'flag';
-            };
-        } else {
-            type = undefined;
-        };
-        return ({ type })
-    };
-    
-    static _getKeyVal(string) {
-        const key = string.match(/(--[^ \n]+)|(-(?:([^-\s])+))/gim)[0];
-        const value = string.slice(key.length + 1).trim();
-        return ({ key: key.trim() == '' ? undefined : key, value: value.trim() == '' ? undefined : value });
-    };
-    
-    static _sanitize(item, type) {
-        switch(type) {
-            case 'key':
-                return (typeof item == 'string') ? item.replace(/^-+/g, '') : item;
-                break;
-            case 'value':
-                return ((typeof item == 'string') && item.match(/(?:^"[^]+"$)|(?:^'[^]+'$)/gim)) ? item.slice(1, -1) : item;
-                break;
-            default:
-                return item;
                 break;
         };
     };
