+++ src/util/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+    identify: require('./identify.js'),
+    sanitize: require('./sanitize.js'),
+    getKeyVal: require('./getKeyVal.js'),
+    typeCoerce: require('./typeCoerce.js'),
+};
